pyproject.toml
[project]
name = "CryptBot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "cam", email = "cammclain@skiff.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.4",
    "sqlalchemy>=2.0.31",
    "logging>=0.4.9.6",
    "outline>=4.0.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "apache-2.0"}


[tool.pdm]
distribution = false
